Make libdiscovercommon plasma-discover-private
The libdiscovercommon name is still very close to the existing
libdiscover name from src:discover, also, since it doesn't contain a
public shared library, keeping the lib* prefix doesn't make much sense.
With the new name it's clear what this package is meant for, and it has
better chances to be accepted by the ftpmasters.